+++ app/Models/SeedTest.php
@@ -2,30 +2,13 @@
 
 namespace App\Models;
 
-use App\Traits\FiltersRecordsByFields;
-use App\Traits\LoadsRelationsThroughServices;
-use App\Traits\OrdersQueryResults;
 use App\Traits\Searchable;
-use App\Traits\Uuids;
 use Illuminate\Database\Eloquent\Factories\HasFactory;
-use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Eloquent\Relations\HasMany;
 
 class SeedTest extends Model
 {
-    use FiltersRecordsByFields;
     use HasFactory;
-    use LoadsRelationsThroughServices;
-    use OrdersQueryResults;
-    use Searchable;
-    use Uuids;
-
-    /**
-     * Indicates if the IDs are auto-incrementing.
-     *
-     * @var bool
-     */
-    public $incrementing = false;
 
     /**
      * The table associated with the model.
@@ -33,20 +16,6 @@ class SeedTest extends Model
      * @var string
      */
     protected $table = 'seed_tests';
-
-    /**
-     * The primary key for the model.
-     *
-     * @var string
-     */
-    protected $primaryKey = 'id';
-
-    /**
-     * The 'type' of the primary key ID.
-     *
-     * @var string
-     */
-    protected $keyType = 'string';
 
     /**
      * The default field used to order query results by.
@@ -63,13 +32,6 @@ class SeedTest extends Model
     protected $orderByDirection = 'asc';
 
     /**
-     * The attributes that are mass assignable.
-     *
-     * @var array
-     */
-    protected $fillable = [];
-
-    /**
      * The attributes that should be visible in serialization.
      *
      * @var array
@@ -81,24 +43,13 @@ class SeedTest extends Model
     ];
 
     /**
-     * The attributes that should be hidden for arrays.
-     *
-     * @var array
-     */
-    protected $hidden = [
-        'created_at',
-        'updated_at',
-        'deleted_at',
-    ];
-
-    /**
      * The attributes that should be cast to native types.
      *
      * @var array
      */
     protected $casts = [
         'id'    => 'string',
-        'level' => 'int',
+        'level' => 'integer',
     ];
 
     /**
