+++ tests/Unit/Models/SeedTestTest.php
@@ -3,25 +3,12 @@
 namespace Tests\Unit\Models;
 
 use App\Models\SeedTest;
-use App\Traits\FiltersRecordsByFields;
-use App\Traits\LoadsRelationsThroughServices;
-use App\Traits\OrdersQueryResults;
-use App\Traits\Searchable;
-use App\Traits\Uuids;
 use Illuminate\Foundation\Testing\RefreshDatabase;
 use Tests\TestCase as TestCase;
 
 class SeedTestTest extends TestCase
 {
     use RefreshDatabase;
-
-    /** @test */
-    public function it_explicitly_disables_incrementing_primary_keys()
-    {
-        $seedTest = new SeedTest();
-
-        $this->assertFalse($seedTest->getIncrementing());
-    }
 
     /** @test */
     public function it_uses_the_proper_database_table()
@@ -32,43 +19,11 @@ class SeedTestTest extends TestCase
     }
 
     /** @test */
-    public function it_sets_the_primary_key_column_explicitly()
-    {
-        $seedTest = new SeedTest();
-
-        $this->assertEquals('id', $seedTest->getKeyName());
-    }
-
-    /** @test */
-    public function it_sets_the_primary_key_type_to_string()
-    {
-        $seedTest = new SeedTest();
-
-        $this->assertEquals('string', $seedTest->getKeyType());
-    }
-
-    /** @test */
     public function it_explicitly_defines_the_column_that_results_should_use_for_ordering()
     {
         $seedTest = new SeedTest();
 
         $this->assertEquals('level', $seedTest->getOrderByField());
-    }
-
-    /** @test */
-    public function it_explicitly_defines_the_order_results_should_be_returned_by()
-    {
-        $seedTest = new SeedTest();
-
-        $this->assertEquals('asc', $seedTest->getOrderByDirection());
-    }
-
-    /** @test */
-    public function it_does_not_allow_properties_to_be_assigned_in_mass()
-    {
-        $seedTest = new SeedTest();
-
-        $this->assertEquals([], $seedTest->getFillable());
     }
 
     /** @test */
@@ -86,20 +41,6 @@ class SeedTestTest extends TestCase
     }
 
     /** @test */
-    public function it_explicitly_defines_the_hidden_fields_for_api_consumption()
-    {
-        $seedTest = new SeedTest();
-
-        $hiddenFields = [
-            'created_at',
-            'updated_at',
-            'deleted_at',
-        ];
-
-        $this->assertEquals($hiddenFields, $seedTest->getHidden());
-    }
-
-    /** @test */
     public function it_explicitly_defines_the_cast_type_for_each_field()
     {
         $seedTest = new SeedTest();
@@ -107,7 +48,7 @@ class SeedTestTest extends TestCase
 
         $expected = [
             'id'    => 'string',
-            'level' => 'int',
+            'level' => 'integer',
         ];
 
         $this->assertEquals($expected, $fields);
@@ -165,60 +106,5 @@ class SeedTestTest extends TestCase
         $seedTest = new SeedTest();
 
         $this->assertEquals('level', $seedTest->getRouteKeyName());
-    }
-
-    /** @test */
-    public function it_uses_uuids_for_the_primary_key()
-    {
-        $this->assertTrue(in_array(
-            Uuids::class,
-            class_uses(SeedTest::class)
-        ));
-    }
-
-    /** @test */
-    public function it_will_not_allow_the_uuid_to_be_changed()
-    {
-        $seedTest = SeedTest::factory()->create();
-
-        $seedTest->id = 'not-original-value';
-        $seedTest->save();
-
-        $this->assertFalse($seedTest->id === 'not-original-value');
-    }
-
-    /** @test */
-    public function it_can_order_query_results()
-    {
-        $this->assertTrue(in_array(
-            OrdersQueryResults::class,
-            class_uses(SeedTest::class)
-        ));
-    }
-
-    /** @test */
-    public function it_includes_search_functionality()
-    {
-        $this->assertTrue(in_array(
-            Searchable::class,
-            class_uses(SeedTest::class)
-        ));
-    }
-
-    /** @test */
-    public function it_includes_the_ability_to_filter_records_by_fields()
-    {
-        $this->assertTrue(in_array(
-            FiltersRecordsByFields::class,
-            class_uses(SeedTest::class)
-        ));
-    }
-
-    public function it_loads_relations_through_services()
-    {
-        $this->assertTrue(in_array(
-            LoadsRelationsThroughServices::class,
-            class_uses(SeedTest::class)
-        ));
     }
 }
